Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    int a,b,c,sum;
+    float a,b,c,sum;
     printf("Average: ");
-    scanf("%d %d %d", &a,&b,&c);
+    scanf("%f %f %f", &a,&b,&c);
     sum=a+b+c;
     float x ;
     x=(sum/3);
